Bug fixes, sigscan system for getting engine ptrs
All seems to be good except for ProcessEvent. It seems to be detoured
properly (and is called properly) but when calling it through
DetourManager::pProcessEvent it seems to run into trouble. The fact that
it's being called is a bit confusing, perhaps the UObject instance isn't
being pushed onto the stack from ecx.