Util library, CrashRpt is now working
Getting CrashRpt to work was quite an ordeal. Apparently the Unreal
basically wraps the entire engine in a __try __except which pretty much
overrides the CrashRpt handlers. So I have to sigscan for the Unreal
exception handler, detour it, and call CrashRpt manually from there in
order for it to work.