Started to implement inheritance system
Basically, if you want to access things directly (eg.
obj.UProperty.Offset) then you can, and probably won't have any
metamethods called, so that will still be pretty efficient. But if you
don't know the class a field belongs to (or can't be bothered typing
it), you can just do obj.Offset and the metamethod on "obj" will cast
"obj" to its actual type (ie. if you have it as a UObject* but it's
actually a UIntProperty* then it will resolve that) and iterate over all
the bases to find the "Offset" field. Right now this only works for
fields, still need a way to do functions.