UObjectIndex doesn't work when it's JIT compiled
I get a "NYI: unsupported C type conversion" in the loop that iterates
over the base classes which means that it just returns nil. When I do
jit.off(), the problem doesn't exist. This is kind of a shame because
it's probably the function that's going to be called the most, so I'll
probably have to get to the bottom of the problem (I suspect there's
nothing that can be done, my method is super hacky and is probably the
reason it's broken).